ETH TX checksum offload (#2916)
- DOL test for ETH TX checksum offload
- Change P4 to insert checksum when gso_valid bit is set
- Change P4+ to set gso_offset as absolute offset instead of being
  relative to gso_start.

+++ dol/config/resmgr.py
@@ -180,9 +180,7 @@ class HostMemory(object):
         memset(va, 0, c_uint64(size))
 
     def __del__(self):
-        if GlobalOptions == None:
-            return
-        if GlobalOptions.dryrun or not GlobalOptions.hostmem: return
+        if not GlobalOptions or GlobalOptions.dryrun or not GlobalOptions.hostmem: return
         self.lib.delete_host_mem()
 
 def init():
